+++ libs/widget-lib/jest.config.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 export default {
   displayName: 'widget-lib',
   preset: '../../jest.preset.js',