+++ libs/iframe-transport/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [2.0.0](https://github.com/cowprotocol/cowswap/compare/iframe-transport-v1.2.0...iframe-transport-v2.0.0) (2026-01-28)
+
+
+### ⚠ BREAKING CHANGES
+
+* migrate from yarn to pnpm ([#6847](https://github.com/cowprotocol/cowswap/issues/6847))
+
+### Features
+
+* migrate from yarn to pnpm ([#6847](https://github.com/cowprotocol/cowswap/issues/6847)) ([1c5d115](https://github.com/cowprotocol/cowswap/commit/1c5d11516d0ddb997b0ebbb4eee4adcb491204a4))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/types bumped to 3.0.0
+
+## [1.2.0](https://github.com/cowprotocol/cowswap/compare/iframe-transport-v1.1.1...iframe-transport-v1.2.0) (2026-01-26)
+
+
+### Features
+
+* fix types in iframe-transport lib ([#6882](https://github.com/cowprotocol/cowswap/issues/6882)) ([f419e21](https://github.com/cowprotocol/cowswap/commit/f419e219fefdd007101ba61053dcce6c218eeee0))
+
 ## [1.1.1](https://github.com/cowprotocol/cowswap/compare/iframe-transport-v1.1.0...iframe-transport-v1.1.1) (2026-01-26)